src/djangoSrc/requirements.txt
google
google-cloud-language
google-cloud-speech
google-cloud-translate
google-api-python-client
pydub
SpeechRecognition
google-cloud-storage
django
djangorestframework
django-cors-headers
mysqlclient
six
pymysql
dropbox